Rework coco project module
The checkout logic wasn't necessarily part of the proxy project, so I
decided to move it to the recently added coco::project module. To
further separate concerns, I split this module into submodules of
checkout and create. This required some refactoring of imports and at
the same time I refactored the errors into separate enums.
